requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pact-python==1.7.0
pluggy==1.0.0
psutil==5.9.5
pydantic==1.10.8
pytest==7.3.1
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
